Move BootCommand to config-api and put back preboot command
